+++ src/runtime_src/xdp/profile/database/events/user_events.cpp
@@ -24,7 +24,7 @@ namespace xdp {
   // User event definitions
   // **************************
 
-  UserMarker::UserMarker(uint64_t s_id, uint64_t ts, uint64_t l) : 
+  UserMarker::UserMarker(uint64_t s_id, double ts, uint64_t l) : 
     VTFEvent(s_id, ts, USER_MARKER), label(l)
   {
   }
@@ -40,7 +40,7 @@ namespace xdp {
     fout << std::endl ;
   }
 
-  UserRange::UserRange(uint64_t s_id, uint64_t ts, bool s, 
+  UserRange::UserRange(uint64_t s_id, double ts, bool s, 
 		       uint64_t l, uint64_t tt) :
     VTFEvent(s_id, ts, USER_RANGE), isStart(s), label(l), tooltip(tt)
   {
